requirements.txt
telethon>=1.36.0
python-dotenv>=1.0.0
tomli>=2.0.0; python_version < "3.11"
